feat(TMST-683): moved date formater to new file
