+++ packages/ts-newskit/src/components/puzzles/puzzle-card/index.tsx
@@ -11,41 +11,15 @@ import {
 import { Wrap, StyledNewsKitPuzzlePlaceholder } from './styles';
 import { PuzzlesFlag } from '../flag';
 import { Puzzle } from '../archive/types';
+import { formatDate } from '../../../utils/date-formatting';
 
 export interface PuzzleCardProps {
   data: Puzzle;
 }
 
 export const PuzzleCard = ({ data }: PuzzleCardProps) => {
-  const timestamp = new Date(data.publishedAt);
-  const cardDate = timestamp.toLocaleString('en-us', {
-    weekday: 'long',
-    month: 'long',
-    day: 'numeric'
-  });
-  const currentDate = new Date();
-  const currentDayOfWeek = currentDate.toLocaleString('en-us', {
-    weekday: 'long',
-    month: 'long',
-    day: 'numeric'
-  });
-  const yesterdayDate = new Date();
-  yesterdayDate.setDate(yesterdayDate.getDate() - 1);
-  const yesterdayDayOfWeek = yesterdayDate.toLocaleString('en-us', {
-    weekday: 'long',
-    month: 'long',
-    day: 'numeric'
-  });
-  let outputDate;
+  const outputDate = formatDate(data.publishedAt);
   const imageUrl = data.image ? data.image.crops[0].url : '';
-
-  if (cardDate === currentDayOfWeek) {
-    outputDate = 'Today';
-  } else if (cardDate === yesterdayDayOfWeek) {
-    outputDate = 'Yesterday';
-  } else {
-    outputDate = cardDate;
-  }
 
   return (
     <CardComposable
